# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tugas LKM 01 PAA",
        "description": "Dokumentasi endpoint API Tugas LKM 01",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current origin server"
        }
    ],
    "paths": {
...
        "/api/borrowings": {
            "get": {
                "tags": [
                    "Borrowings"
                ],
                "summary": "Get all borrowings",
                "operationId": "d1ddcecfcc1954a3a1abe1a386242e0d",
                "responses": {
                    "200": {
                        "description": "Borrowings retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Borrowings"
                ],
                "summary": "Create a new borrowing",
                "operationId": "6b92df97f30f4c58a958bdc7d9234656",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "book_id",
                                    "borrower_name",
                                    "borrower_email",
                                    "borrowed_at",
                                    "due_date",
                                    "status"
                                ],
                                "properties": {
                                    "book_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "borrower_name": {
                                        "type": "string",
                                        "example": "Kiarra"
                                    },
                                    "borrower_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "kiarra@example.com"
                                    },
                                    "borrowed_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-16"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-23"
                                    },
                                    "returned_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "borrowed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Borrowing created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/borrowings/{borrowing}": {
            "get": {
                "tags": [
                    "Borrowings"
                ],
                "summary": "Get a borrowing by id",
                "operationId": "889c7f8926a86e808837d1a97e050d64",
                "parameters": [
                    {
                        "name": "borrowing",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Borrowing retrieved successfully"
                    },
                    "404": {
                        "description": "Borrowing not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Borrowings"
                ],
                "summary": "Update a borrowing",
                "operationId": "4f8c42bf3dcda13e1744a2e73bd8c8ad",
                "parameters": [
                    {
                        "name": "borrowing",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "book_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "borrower_name": {
                                        "type": "string",
                                        "example": "Kiarra"
                                    },
                                    "borrower_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "kiarra@example.com"
                                    },
                                    "borrowed_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-16"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-23"
                                    },
                                    "returned_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "returned"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Borrowing updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Borrowings"
                ],
                "summary": "Delete a borrowing",
                "operationId": "30795a32d6e714f77dbf13dcd3d718f9",
                "parameters": [
                    {
                        "name": "borrowing",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Borrowing deleted successfully"
                    },
                    "404": {
                        "description": "Borrowing not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT dengan format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}